moved test messages into fixtures class. Minor refactoring of messages case statements.
