+++ test/v1/services/GetCalculationServiceSpec.scala
@@ -27,6 +27,7 @@ import v1.models.outcomes.ResponseWrapper
 import v1.models.request.getCalculation.GetCalculationRequest
 import v1.models.response.common.{Message, Messages, Metadata}
 import v1.models.response.getCalculation.GetCalculationResponse
+import v1.fixtures.Fixtures._
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
@@ -46,12 +47,6 @@ class GetCalculationServiceSpec extends UnitSpec {
     calculationErrorCount = Some(1)
   )
 
-  val err1 = Message("err1", "text1")
-  val err2 = Message("err2", "text2")
-  val info1 = Message("info1", "text1")
-  val info2 = Message("info2", "text2")
-  val warn1 = Message("warn1", "text1")
-  val warn2 = Message("warn2", "text2")
   val messagesResponse = Messages(Some(Seq(info1,info2)), Some(Seq(warn1,warn2)), Some(Seq(err1,err2)))
 
   val getCalculationResponse = GetCalculationResponse(metadataResponse, messages = Some(messagesResponse))
